requirements.txt
flask==2.3.3
llama-cpp-python>=0.2.0
gunicorn==23.0.0
requests>=2.31.0